AGENT_1/AGENT_1/requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
pandas==2.0.3
requests==2.31.0
textstat==0.7.3
chromadb==0.4.22
crawl4ai==0.4.2